% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coastr.R
\docType{package}
\name{coastr}
\alias{coastr}
\title{coastr: Tools for coastal analysis using R}
\description{
The coastr package provides functions for coastal analysis,
including beach profile analysis generated from high resolution
digital surface model, time-series analysis of shoreline positions.
}
\section{coastr functions}{

}

