% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.rad.R
\name{deg.rad}
\alias{deg.rad}
\title{A degree conversion function}
\usage{
deg.rad(deg)
}
\arguments{
\item{deg}{Degree unit to be converted to radian.}
}
\description{
This function converts degrees to radians.
}
\examples{
deg.rad(360)
}
\keyword{degree}
\keyword{units}
