% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2rad.R
\name{deg2rad}
\alias{deg2rad}
\title{A degree conversion function}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{Degree unit to be converted to radian.}
}
\description{
This function converts degrees to radians.
}
\examples{
deg2rad(360)
}
\keyword{degree}
\keyword{units}
