% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_rad.R
\name{deg_rad}
\alias{deg_rad}
\title{A degree conversion function}
\usage{
deg_rad(deg)
}
\arguments{
\item{deg}{Degree unit to be converted to radian.}
}
\description{
This function converts degrees to radians.
}
\examples{
deg_rad(360)
}
\keyword{degree}
\keyword{units}
