% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_sites.R
\name{seq_sites}
\alias{seq_sites}
\title{Sequential coastal site order}
\usage{
seq_sites(site_list, reverse = FALSE, coast = TRUE)
}
\arguments{
\item{site_list}{A dataframe with a 'lon' and 'lat' column that
contains coordinates along a coastline that need to be ordered.}

\item{reverse}{A boolean TRUE/FALSE to determine if the ordering
of the sites should be reversed on output. Default = FALSE.}

\item{coast}{Should only the coastline be used to calculate the
sequence of sites? Or should islands be included? Note that
sites located on islands will have their sequence of order
calculated after sites along the coast. This is generally not a
desired outcome. Default = TRUE.}
}
\value{
A dataframe with the correct ordering of the input sites
}
\description{
This function finds the order of sites along a convoluted but
contiguous coastline. It is not designed to work for multiple
different coastlines at once.
}
\examples{
SACTN_site_list_sub <- SACTN_site_list[17:32,]
SACTN_site_list_ordered <- seq_sites(SACTN_site_list_sub)
}
\keyword{sequential}
\keyword{sites}
