% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{transects}
\alias{transects}
\title{Coastal Transects}
\usage{
transects(site_list, alongshore = FALSE, spread = 2, coast = TRUE,
  reverse = FALSE)
}
\arguments{
\item{site_list}{A dataframe with a 'lon' and 'lat' column that
contains coordinates for sites along a coastline.}

\item{alongshore}{A boolean TRUE/FALSE to determine if the
transects should be shore-normal or alongshore. Default = FALSE.}

\item{spread}{The number of pixels along the coast on either
side of the site in question that should be used to determine
the angle of the shore-normal transect. A larger spread value
will provide smoother angles, which may be undesirable for
convoluted coastlines. Defaut = 2.}

\item{coast}{Should only the coastline be used to calculate the
angle for transects? Or should islands be included? Calculating
the angle of transects form or along the faces of islands is
advised against. The coastline behind the island should almost
always be used instead. Default = TRUE.}

\item{reverse}{A boolean TRUE/FALSE that will reverse the angle
of the shore-normal transects. Default = FALSE.}
}
\value{
A dataframe with the correct ordering of the input sites
}
\description{
This function finds the correct angle for transects that
run either along or away from a set of points along a coastline.
}
\examples{
SACTN_site_list_transects <- transects(SACTN_site_list[25:45,])
}
\keyword{alongshore}
\keyword{shore-normal}
\keyword{transects}
