% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coastR-package.R
\docType{package}
\name{coastR}
\alias{coastR}
\alias{coastR-package}
\title{coastR.}
\description{
This package contains several functions designed to address common problems
faced in coastal oceanography.
}
\details{
The main function in this package currently is \code{\link{seqSites}}.
This function allows one to automagically organise sites along a stretch of
coastline when the coast is too convoluted to do so simply using longitude
or latitude.

There are a few other small functions included and more will be added in time.
}
\references{
Schlegel, R.W., Smit, A.J., 2016. Climate change in coastal waters:
time series properties affecting trend estimation. J. Clim. 29 (24), 9113–9124.
<DOI:10.1175/JCLI-D-16-0014.1>
}
\author{
Robert W. Schlegel <\email{wiederweiter@gmail.com}>
}
