% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-transect.R
\name{create_transect}
\alias{create_transect}
\title{Create transects}
\usage{
create_transect(
  x,
  transect_spacing = NULL,
  length_left = if (all(!is.null(transect_spacing), is.numeric(transect_spacing)))
    transect_spacing,
  length_right = if (all(!is.null(transect_spacing), is.numeric(transect_spacing)))
    transect_spacing,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{x}{name of an object of class \code{sf}, or a path to a file
readable by the \code{sf::st_read} function, which must contain
geometries of type \code{LINESTRING} or \code{MULTILINESTRING}.}

\item{transect_spacing}{character or numeric. If numeric, represents
the spacing between transects in \code{x} CRS units. If character,
accepts only the string 'vertices' as argument (allows partial matching)}

\item{length_left}{numeric; length of transects to the right of the input line}

\item{reverse}{logical; reverse orientation of
transects; default \code{FALSE}}

\item{...}{other arguments passed to sf::st_read.}
}
\value{

}
\description{
Create transects of user-defined length along an input line
}
\details{

}
\examples{

}
