% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-transect.R
\name{create_transect}
\alias{create_transect}
\title{Create transects}
\usage{
create_transect(obj, distance = NULL, transect_length = 1, sides)
}
\arguments{
\item{distance}{numeric; separation distance between transects}

\item{transect_length}{numeric; Euclidean distance of transects}

\item{sides}{character; takes 'left', 'right' and 'both'}

\item{x}{name of an object of class sf which must contain
geometries of type \code{LINESTRING} or \code{MULTILINESTRING}.}
}
\description{
Create transects of user-defined length and at regular distance
along an input shoreline
}
