% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.transect}
\title{Plot transect object}
\usage{
\method{plot}{transect}(x, ...)
}
\arguments{
\item{x}{object of class transect}

\item{...}{other arguments passed to ggplot2::geom_sf}
}
\value{
An object of class ggplot.
}
\description{
Plot transects with along with basic attributes
}
\details{
plot.transect plots the transect using a minimal theme,
  and place a label with the ID and attributes over its centroids.
}
\examples{

(file_path <- system.file('extdata', 'transects-long.geojson', package = 'coastr'))
(transect <- read_transect(file_path, profile = T, start_from = 'N'))
plot(transect)

}
